Create Find the closest pair from two arrays
Given two sorted arrays arr[] and brr[] and a number x, find the pair whose sum is closest to x and the pair has an element from each array. In the case of multiple closest pairs return any one of them. Note: Can return the two numbers in any manner.